backend/auth.py
from jose import jwt
from bcrypt import checkpw, hashpw, gensalt
import os

# JWT ayarları
SECRET_KEY = os.getenv("JWT_SECRET_KEY", "your-secret-key")
ALGORITHM = "HS256"

# Password hashing
def verify_password(plain_password, hashed_password):
    try:
        return checkpw(plain_password.encode('utf-8'), hashed_password.encode('utf-8'))
    except Exception as e:
        print(f"Password verification error: {e}")
        return False

def get_password_hash(password):
    return hashpw(password.encode('utf-8'), gensalt()).decode('utf-8')

def create_access_token(data: dict):
    try:
        # Token'ı süresiz olarak oluştur
        to_encode = data.copy()
        encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
        return encoded_jwt
    except Exception as e:
        print(f"Token creation error: {e}")
        raise e 