backend/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-jose[cryptography]==3.4.0
bcrypt==4.3.0
python-multipart==0.0.20
pydantic[email]==2.1.1
requests==2.32.3
tensorflow==2.19.0
numpy==1.26.4
Pillow==11.1.0